feat: finalize addRelativeTo() implement; add README
